fastapi-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
motor>=3.6.0,<4
pydantic[email]==2.7.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
dnspython==2.6.1
